move the app to external and with necessary changes
