+++ firmware/application/apps/ui_external_module_view.cpp
@@ -92,6 +92,9 @@ void ExternalModuleView::on_tick_second() {
             case app_location_t::TX:
                 btnText += " (TX)";
                 break;
+            case app_location_t::SETTINGS:
+                btnText += " (Settings)";
+                break;
             case app_location_t::DEBUG:
                 btnText += " (Debug)";
                 break;
