+++ firmware/application/apps/ui_settings.cpp
@@ -1094,243 +1094,6 @@ void SetBatteryView::focus() {
     button_cancel.focus();
 }
 
-/* AppManagerView **************************************/
-/* | inner apps | external apps |
- * | id         | display name  |
- * | callname   | friendly name |
- */
-AppManagerView::AppManagerView(NavigationView& nav)
-    : nav_{nav} {
-    add_children({&labels,
-                  &menu_view,
-                  &text_app_info,
-                  &button_hide_unhide,
-                  &button_clean_hide,
-                  &button_set_cancel_autostart,
-                  &button_clean_autostart});
-
-    menu_view.set_parent_rect({0, 2 * 8, screen_width, 24 * 8});
-
-    menu_view.on_highlight = [this]() {
-        if (menu_view.highlighted_index() >= app_list_index) {
-            text_app_info.set("");
-            return;
-        }
-
-        std::string info;
-        auto app_name = get_app_info(menu_view.highlighted_index(), true);
-        auto app_id = get_app_info(menu_view.highlighted_index(), false);
-
-        info += "Hidden:";
-
-        if (is_blacklisted(app_name)) {
-            info += "Yes ";
-        } else {
-            info += "No ";
-        }
-
-        info += "Autostart:";
-        if (is_autostart_app(app_id)) {
-            info += "Yes";
-        } else {
-            info += "No";
-        }
-
-        if (info.empty()) {
-            info = "Highlight an app";
-        }
-
-        text_app_info.set(info);
-    };
-
-    button_hide_unhide.on_select = [this](Button&) {
-        hide_unhide_app();
-        refresh_list();
-    };
-
-    button_clean_hide.on_select = [this](Button&) {
-        clean_blacklist();
-        refresh_list();
-    };
-
-    button_set_cancel_autostart.on_select = [this](Button&) {
-        set_unset_autostart_app();
-        refresh_list();
-    };
-
-    button_clean_autostart.on_select = [this](Button&) {
-        unset_auto_start();
-        refresh_list();
-    };
-
-    refresh_list();
-}
-
-void AppManagerView::refresh_list() {
-    auto previous_cursor_index = menu_view.highlighted_index();
-    app_list_index = 0;
-    menu_view.clear();
-
-    size_t index = 0;
-
-    auto padding = [](const std::string& str) {
-        return str.length() < 25 ? std::string(25 - str.length(), ' ') + '' : "";
-        // that weird char is a icon in portapack's font base that looks like a switch, which i use to indicate the auto start
-    };
-
-    for (auto& app : NavigationView::appList) {
-        if (app.id == nullptr) continue;
-
-        app_list_index++;
-
-        menu_view.add_item({app.displayName + std::string(is_autostart_app(app.id) ? padding(app.displayName) : ""),
-                            app.iconColor,
-                            app.icon,
-                            [this, app_id = std::string(app.id)](KeyEvent) {
-                                button_hide_unhide.focus();
-                            }});
-    }
-
-    ExternalItemsMenuLoader::load_all_external_items_callback([this, &index, &padding](ui::AppInfoConsole& app) {
-        if (app.appCallName == nullptr) return;
-
-        app_list_index++;
-
-        menu_view.add_item({app.appFriendlyName + std::string(is_autostart_app(app.appCallName) ? padding(app.appFriendlyName) : ""),
-                            ui::Color::light_grey(),
-                            &bitmap_icon_sdcard,
-                            [this, app_id = std::string(app.appCallName)](KeyEvent) {
-                                button_hide_unhide.focus();
-                            }});
-    });
-
-    menu_view.set_highlighted(previous_cursor_index);
-}
-
-void AppManagerView::focus() {
-    menu_view.focus();
-}
-
-void AppManagerView::hide_app() {
-    std::vector<std::string> blacklist;
-    get_blacklist(blacklist);
-
-    blacklist.push_back(get_app_info(menu_view.highlighted_index(), true));
-    write_blacklist(blacklist);
-}
-
-void AppManagerView::unhide_app() {
-    std::vector<std::string> blacklist;
-    get_blacklist(blacklist);
-    blacklist.erase(std::remove(blacklist.begin(), blacklist.end(), get_app_info(menu_view.highlighted_index(), true)), blacklist.end());
-    write_blacklist(blacklist);
-}
-
-void AppManagerView::hide_unhide_app() {
-    if (is_blacklisted(get_app_info(menu_view.highlighted_index(), true)))
-        unhide_app();
-    else
-        hide_app();
-}
-
-void AppManagerView::get_blacklist(std::vector<std::string>& blacklist) {
-    File f;
-
-    auto error = f.open(u"SETTINGS/blacklist");
-    if (error)
-        return;
-
-    auto reader = FileLineReader(f);
-    for (const auto& line : reader) {
-        if (line.length() == 0 || line[0] == '#')
-            continue;
-
-        blacklist.push_back(trim(line));
-    }
-}
-
-void AppManagerView::write_blacklist(const std::vector<std::string>& blacklist) {
-    File f;
-    auto error = f.create(u"SETTINGS/blacklist");
-    if (error)
-        return;
-
-    for (const auto& app_name : blacklist) {
-        auto line = app_name + "\n";
-        f.write(line.c_str(), line.length());
-    }
-}
-
-void AppManagerView::clean_blacklist() {
-    std::vector<std::string> blacklist = {};
-    write_blacklist(blacklist);
-}
-
-bool AppManagerView::is_blacklisted(const std::string& app_name) {
-    std::vector<std::string> blacklist;
-    get_blacklist(blacklist);
-    return std::find(blacklist.begin(), blacklist.end(), app_name) != blacklist.end();
-}
-
-void AppManagerView::set_auto_start() {
-    auto app_index = menu_view.highlighted_index();
-    if (app_index >= app_list_index) return;
-
-    auto id_aka_friendly_name = get_app_info(app_index, false);
-
-    autostart_app = id_aka_friendly_name;
-
-    refresh_list();
-}
-
-void AppManagerView::unset_auto_start() {
-    autostart_app = "";
-    refresh_list();
-}
-
-void AppManagerView::set_unset_autostart_app() {
-    auto app_index = menu_view.highlighted_index();
-    if (app_index >= app_list_index) return;
-
-    auto id_aka_friendly_name = get_app_info(app_index, false);
-
-    if (is_autostart_app(id_aka_friendly_name))
-        unset_auto_start();
-    else
-        set_auto_start();
-}
-
-bool AppManagerView::is_autostart_app(const char* id_aka_friendly_name) {
-    return autostart_app == std::string(id_aka_friendly_name);
-}
-
-bool AppManagerView::is_autostart_app(const std::string& id_aka_friendly_name) {
-    return autostart_app == id_aka_friendly_name;
-}
-
-std::string AppManagerView::get_app_info(uint16_t index, bool is_display_name) {
-    size_t current_index = 0;
-    std::string result;
-
-    for (auto& app : NavigationView::appList) {
-        if (app.id == nullptr) continue;
-        if (current_index == index) {
-            return is_display_name ? std::string(app.displayName) : std::string(app.id);
-        }
-        current_index++;
-    }
-
-    ExternalItemsMenuLoader::load_all_external_items_callback([&current_index, &index, &result, &is_display_name](ui::AppInfoConsole& app) {
-        if (app.appCallName == nullptr) return;
-        if (current_index == index) {
-            result = is_display_name ? app.appFriendlyName : app.appCallName;
-        }
-        current_index++;
-    });
-
-    return result;
-}
-
 /* SettingsMenuView **************************************/
 
 SettingsMenuView::SettingsMenuView(NavigationView& nav)
@@ -1339,11 +1102,13 @@ SettingsMenuView::SettingsMenuView(NavigationView& nav)
 }
 
 void SettingsMenuView::on_populate() {
-    if (pmem::show_gui_return_icon()) {
+    const bool return_icon = pmem::show_gui_return_icon();
+
+    if (return_icon) {
         add_items({{"..", ui::Color::light_grey(), &bitmap_icon_previous, [this]() { nav_.pop(); }}});
     }
+
     add_items({
-        {"App Manager", ui::Color::dark_cyan(), &bitmap_icon_notepad, [this]() { nav_.push<AppManagerView>(); }},
         {"App Settings", ui::Color::dark_cyan(), &bitmap_icon_notepad, [this]() { nav_.push<AppSettingsView>(); }},
         {"Audio", ui::Color::dark_cyan(), &bitmap_icon_speaker, [this]() { nav_.push<SetAudioView>(); }},
         {"Calibration", ui::Color::dark_cyan(), &bitmap_icon_options_touch, [this]() { nav_.push<TouchCalibrationView>(); }},
@@ -1363,7 +1128,10 @@ void SettingsMenuView::on_populate() {
         {"Theme", ui::Color::dark_cyan(), &bitmap_icon_setup, [this]() { nav_.push<SetThemeView>(); }},
         {"Autostart", ui::Color::dark_cyan(), &bitmap_icon_setup, [this]() { nav_.push<SetAutostartView>(); }},
     });
+
     if (battery::BatteryManagement::isDetected()) add_item({"Battery", ui::Color::dark_cyan(), &bitmap_icon_batt_icon, [this]() { nav_.push<SetBatteryView>(); }});
+
+    add_external_items(nav_, app_location_t::SETTINGS, *this, return_icon ? 1 : 0);
 }
 
 } /* namespace ui */
