+++ firmware/application/external/external.ld
@@ -58,6 +58,7 @@ MEMORY
     ram_external_app_fmradio(rwx) : org = 0xADD10000, len = 32k 
     ram_external_app_tuner(rwx) : org = 0xADD20000, len = 32k
     ram_external_app_metronome(rwx) : org = 0xADD30000, len = 32k 
+    ram_external_app_app_manager(rwx) : org = 0xADD40000, len = 32k 
 }
 
 SECTIONS
@@ -272,4 +273,10 @@ SECTIONS
         KEEP(*(.external_app.app_metronome.application_information));
         *(*ui*external_app*metronome*);
     } > ram_external_app_metronome
+
+    .external_app_app_manager : ALIGN(4) SUBALIGN(4)
+    {
+        KEEP(*(.external_app.app_app_manager.application_information));
+        *(*ui*external_app*app_manager*);
+    } > ram_external_app_app_manager
 }
