+++ firmware/common/standalone_app.hpp
@@ -71,7 +71,8 @@ enum app_location_t : uint32_t {
     RX,
     TX,
     DEBUG,
-    HOME
+    HOME,
+    SETTINGS
 };
 
 struct standalone_application_information_t {
